+++ Cargo.toml
@@ -1,10 +1,6 @@
 [package]
 name = "ellington"
-<<<<<<< HEAD
 version = "0.2.0"
-=======
-version = "0.1.6"
->>>>>>> master
 authors = ["Adam Harries <harries.adam@gmail.com>"]
 build = "build.rs"
 readme = "README.md"
